requirements.txt
# YuNi 项目依赖
# 生产环境: pip install -r requirements.txt
# 开发环境: pip install -r requirements-dev.txt
# 版本锁定: 使用 requirements-locked.txt 确保可复现性

# ==================== Web框架 ====================
fastapi>=0.100.0,<0.116.0
uvicorn[standard]>=0.20.0,<0.33.0
python-multipart>=0.0.6,<0.1.0
jinja2>=3.1.0,<4.0.0

# ==================== 语音识别 ====================
openai-whisper>=20231117
faster-whisper>=1.0.0,<2.0.0
fast-langdetect>=1.0.0,<2.0.0
split-lang>=2.1.1,<3.0.0
librosa>=0.11.0,<1.0.0
transformers>=4.41.2,<5.0.0
x-transformers>=2.17.3,<3.0.0
peft>=0.18.1,<1.0.0
accelerate>=1.13.0,<2.0.0
sentencepiece>=0.2.1,<1.0.0
ffmpeg-python>=0.2.0,<1.0.0
torchmetrics>=1.9.0,<2.0.0
pytorch-lightning>=2.6.1,<3.0.0
matplotlib>=3.10.8,<4.0.0
cn2an>=0.5.23,<1.0.0
pypinyin>=0.55.0,<1.0.0
omegaconf>=2.3.0,<3.0.0
scipy>=1.17.1,<2.0.0

# ==================== 用户管理系统 ====================
sqlalchemy>=2.0.23,<3.0.0
werkzeug>=2.3.7,<4.0.0
requests>=2.28.0,<3.0.0
PyJWT>=2.8.0,<3.0.0

# ==================== 图片处理 ====================
Pillow>=10.0.0,<12.0.0

# ==================== RAG 服务 ====================
jieba>=0.42.1,<0.43.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0

# ==================== 可选依赖（高级功能）====================
# 取消注释以启用高级功能:
# sentence-transformers>=2.2.0  # 高级语义检索
# faiss-cpu>=1.7.4  # 高效向量检索
# transformers>=4.30.0  # 中文问答模型




